feat: Change database backup and restore API
Now backup files are compressed with gzip by default, and the restore function can automatically decompress and handle both .sql and .sql.gz files.

feat: Change the way to delete website directories

When a website is deleted, its directory will be moved to the /tmp directory for a certain period of time, making it easier to recover.

style: Improve variable names

fix: Fix error when creating applications and trying to access the database